+++ upcloud/provider.go
@@ -167,6 +167,7 @@ func (p *upcloudProvider) Resources(_ context.Context) []func() resource.Resourc
 		loadbalancer.NewLoadBalancerResource,
 		loadbalancer.NewManualCertificateBundleResource,
 		loadbalancer.NewResolverResource,
+		managedobjectstorage.NewManagedObjectStorageBucketResource,
 		managedobjectstorage.NewManagedObjectStorageCustomDomainResource,
 		managedobjectstorage.NewManagedObjectStoragePolicyResource,
 		network.NewNetworkResource,