+++ upcloud/testdata/upcloud_managed_object_storage/managed_object_storage_s1.tf
@@ -61,3 +61,8 @@ resource "upcloud_managed_object_storage" "minimal" {
 
   # No networks or labels
 }
+
+resource "upcloud_managed_object_storage_bucket" "this" {
+  service_uuid = upcloud_managed_object_storage.minimal.id
+  name = "bucket"
+}